# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "api": {},
        "/reg/user": {
            "get": {
                "tags": [
                    "RegistrationAPI"
                ],
                "summary": "Запрос проверки существования пользователя",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "User login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "RegistrationAPI"
                ],
                "summary": "Запрос добавления нового пользователя",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "repeat_password": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "gender_id": {
                                        "type": "boolean"
                                    },
                                    "photo": {
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "login": "login",
                                    "password": "!ReeA178gG",
                                    "repeat_password": "!ReeA178gG",
                                    "phone": "+79999999999",
                                    "email": "email@mail.ru",
                                    "city": "City",
                                    "gender_id": 0,
                                    "photo": []
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully registered user"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}